Remove fixer class and merge with clean_data
